feat(Ristretto): Introduce SetIfPresent and refactor codebase (#282)
* Introduce SetIfPresent and move metrics out to new file.
* Introduce ShouldUpdate, to allow Dgraph to only update the cache with a higher version entry.
* Remove interfaces.
* Rename structs for clarity.
