+++ policy_test.go
@@ -15,14 +15,14 @@ func TestPolicy(t *testing.T) {
 }
 
 func TestPolicyMetrics(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.CollectMetrics(newMetrics())
 	require.NotNil(t, p.metrics)
-	require.NotNil(t, p.evict.metrics)
+	require.NotNil(t, p.costs.metrics)
 }
 
 func TestPolicyProcessItems(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.itemsCh <- []uint64{1, 2, 2}
 	time.Sleep(wait)
 	p.Lock()
@@ -39,7 +39,7 @@ func TestPolicyProcessItems(t *testing.T) {
 }
 
 func TestPolicyPush(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	require.True(t, p.Push([]uint64{}))
 
 	keepCount := 0
@@ -52,12 +52,12 @@ func TestPolicyPush(t *testing.T) {
 }
 
 func TestPolicyAdd(t *testing.T) {
-	p := newDefaultPolicy(1000, 100)
+	p := newPolicy(1000, 100)
 	if victims, added := p.Add(1, 101); victims != nil || added {
 		t.Fatal("can't add an item bigger than entire cache")
 	}
 	p.Lock()
-	p.evict.add(1, 1)
+	p.costs.add(1, 1)
 	p.admit.Increment(1)
 	p.admit.Increment(2)
 	p.admit.Increment(3)
@@ -81,14 +81,14 @@ func TestPolicyAdd(t *testing.T) {
 }
 
 func TestPolicyHas(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.Add(1, 1)
 	require.True(t, p.Has(1))
 	require.False(t, p.Has(2))
 }
 
 func TestPolicyDel(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.Add(1, 1)
 	p.Del(1)
 	p.Del(2)
@@ -97,29 +97,29 @@ func TestPolicyDel(t *testing.T) {
 }
 
 func TestPolicyCap(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.Add(1, 1)
 	require.Equal(t, int64(9), p.Cap())
 }
 
 func TestPolicyUpdate(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.Add(1, 1)
 	p.Update(1, 2)
 	p.Lock()
-	require.Equal(t, int64(2), p.evict.keyCosts[1])
+	require.Equal(t, int64(2), p.costs.keyCosts[1])
 	p.Unlock()
 }
 
 func TestPolicyCost(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.Add(1, 2)
 	require.Equal(t, int64(2), p.Cost(1))
 	require.Equal(t, int64(-1), p.Cost(2))
 }
 
 func TestPolicyClear(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.Add(1, 1)
 	p.Add(2, 2)
 	p.Add(3, 3)
@@ -135,20 +135,20 @@ func TestPolicyClose(t *testing.T) {
 		require.NotNil(t, recover())
 	}()
 
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.Add(1, 1)
 	p.Close()
 	p.itemsCh <- []uint64{1}
 }
 
 func TestPushAfterClose(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.Close()
 	require.False(t, p.Push([]uint64{1, 2}))
 }
 
 func TestAddAfterClose(t *testing.T) {
-	p := newDefaultPolicy(100, 10)
+	p := newPolicy(100, 10)
 	p.Close()
 	p.Add(1, 1)
 }
